+++ Libplanet/Net/Messages/TxIds.cs
@@ -8,25 +8,21 @@ namespace Libplanet.Net.Messages
 {
     internal class TxIds : Message
     {
-        public TxIds(Address sender, IEnumerable<TxId> txIds)
+        public TxIds(IEnumerable<TxId> txIds)
         {
-            Sender = sender;
             Ids = txIds;
         }
 
         public TxIds(byte[][] dataFrames)
         {
-            Sender = new Address(dataFrames[0]);
-            int txCount = BitConverter.ToInt32(dataFrames[1], 0);
+            int txCount = BitConverter.ToInt32(dataFrames[0], 0);
             Ids = dataFrames
-                .Skip(2).Take(txCount)
+                .Skip(1).Take(txCount)
                 .Select(ba => new TxId(ba))
                 .ToList();
         }
 
         public IEnumerable<TxId> Ids { get; }
-
-        public Address Sender { get; }
 
         public override MessageType Type => MessageType.TxIds;
 
@@ -35,7 +31,6 @@ namespace Libplanet.Net.Messages
             get
             {
                 var frames = new List<byte[]>();
-                frames.Add(Sender.ToByteArray());
                 frames.Add(BitConverter.GetBytes(Ids.Count()));
                 frames.AddRange(Ids.Select(id => id.ToByteArray()));
                 return frames;